feat(uses): Use uses for everything that  has them closes #260
