+++ lib/config-ui.js
@@ -780,7 +780,7 @@ class SheetEnhancementsMenu extends ConfigMenu {
         path: `${root}.autoSpellSlots`, title: 'Process Spell Slots Automatically', menuCmd: menu,
       }) +
       this.makeToggleSetting({
-        path: `${root}.autoTraits`, title: 'Process Traits automatically', menuCmd: menu,
+        path: `${root}.autoTraits`, title: 'Process Uses automatically', menuCmd: menu,
       });
 
     const th = utils.buildHTML('th', 'Character Sheet Enhancements', { colspan: '2' });
