+++ lib/shaped-script.js
@@ -12,7 +12,7 @@ const AbilityMaker = require('./ability-maker');
 const ConfigUI = require('./config-ui');
 const AdvantageTracker = require('./advantage-tracker');
 const RestManager = require('./rest-manager');
-const TraitManager = require('./trait-manager');
+const UsesManager = require('./uses-manager');
 const AmmoManager = require('./ammo-manager');
 /**
  * @typedef {Object} ChatMessage
@@ -41,7 +41,7 @@ function ShapedScripts(logger, myState, roll20, parser, entityLookup, reporter, 
   const commandProc = makeCommandProc('shaped', roll20);
   const chatWatchers = [];
   const advantageTracker = new AdvantageTracker();
-  const traitManager = new TraitManager();
+  const usesManager = new UsesManager();
   const ammoManager = new AmmoManager();
   const abilityMaker = new AbilityMaker();
   const errorHandler = function errorHandler(e) {
@@ -59,7 +59,7 @@ function ShapedScripts(logger, myState, roll20, parser, entityLookup, reporter, 
     abilityMaker,
     new ConfigUI(),
     advantageTracker,
-    traitManager,
+    usesManager,
     new RestManager(),
     ammoManager,
   ];
@@ -813,9 +813,9 @@ function ShapedScripts(logger, myState, roll20, parser, entityLookup, reporter, 
     }
   };
 
-  this.handleTraitClick = function handleTraitClick(options) {
+  this.handleUses = function handleTraitClick(options) {
     if (myState.config.sheetEnhancements.autoTraits) {
-      traitManager.handleTraitClick(options);
+      usesManager.handleUses(options);
     }
   };
 
@@ -1175,7 +1175,7 @@ function ShapedScripts(logger, myState, roll20, parser, entityLookup, reporter, 
     this.registerChatWatcher(this.handleHD, ['character', 'title']);
     this.registerChatWatcher(this.handleD20Roll, ['character', '2d20kh1']);
     this.registerChatWatcher(this.handleD20Roll, ['character', '2d20kl1']);
-    this.registerChatWatcher(this.handleTraitClick, ['character', 'trait']);
+    this.registerChatWatcher(this.handleUses, ['character', 'uses', 'repeatingItem']);
     // this.registerChatWatcher(this.handleSpellCast, ['character', 'spell', 'friendlyLevel']);
   };
