+++ package.json
@@ -5,7 +5,7 @@
   "main": "lib/entry-point.js",
   "dependencies": {
     "roll20-logger": "^2.2.0",
-    "roll20-wrapper": "^1.9.0",
+    "roll20-wrapper": "latest",
     "underscore": "^1.8.3"
   },
   "devDependencies": {
